Move StringManager to its own component
Test: build, run StringManagerTests
