+++ CMakeLists.txt
@@ -162,6 +162,7 @@ add_subdirectory(src/OrbitPaths)
 add_subdirectory(src/OrbitTest)
 add_subdirectory(src/OrbitVersion)
 add_subdirectory(src/PresetFile)
+add_subdirectory(src/StringManager)
 add_subdirectory(src/Symbols)
 add_subdirectory(src/TracingInterface)
 add_subdirectory(src/Test)
